requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.12.14
aiofiles==24.1.0
tmdbv3api==1.9.0
yt-dlp==2025.7.21
gunicorn==21.2.0
paramiko==3.4.0
pysmb==1.2.9
python-nmap==0.7.1
cryptography==41.0.7
Pillow==10.1.0
python-magic==0.4.27
watchdog==3.0.0
schedule==1.2.0

